Does not try to open a article using the array itself
Previously, it would go thru each element in the list (as expected)
but it would also try to open an article by using the entire list.
I corrected this behaviour by having it only process if the input
is not an object (lists are objects in javascript).

It works correctly, so I updated from 1.2.0 to 1.2.1