Corrected the autocommit system
Before, it would make a commit after I published to NPM, but it would
not actually add any of the changes made. Now, it will commit everything
after I publish to NPM. This is OK because I only make 1 change at a time
when I publish to NPM to keep it simple.