BUGFIX: Cannot find module './parse.coffee'
This is a response to issue #1, which contains the matching title.
The problem was that the 'parse.coffee' file was not being compiled,
but was nonetheless being required by the compiled 'main.js'

To fix this, Gulp will now compile all the *.coffee source files to
compiled/ and then will not use Browserify to link them together.
Instead, we are just going to use Node's native CommonJS implementation.

The problem with that solution was that the source files required
a file with the .coffee extension e.g. `require("./parse.coffee")`
whereas the compiled versions would all have the .js extension.
To fix this, it now just requires './parse' so that the .js is implicit.
When using the `coffee` environment, it will assume that you want
parse.coffee and it will assume parse.js when run in Node for the compiled.

Also, this slightly makes build times shorter.